Detect start script in package.json
Adds logic to only successfully detect a project with a start script in
its `package.json`.

Extracts logic for parsing a `package.json` and reuses it in the
detection step.

Adds new unit tests to cover the changes and adjust previous tests with
positive detection to use a `package.json` with a start script.

Adjusts integrations tests by removing the simple-app case. According to
this PR that case should no longer be detected.